Cleaned up specialVicinity code
